Initial commit
I just updated the libft from the transfert version of ft_ls. This
one has a generic collector function to get everything that needs
to be cleaned before exit. This not yet tested for leaks, so this
will have to be done. But as for now, it works. The program compiles
and the behaviour is coherent even with demanding tasks (thats why
I wanted to test the new version of the libft with ft_ls).